Merged literals and comments into a single array and object store.
Renamed "data" script property to "lines" to better describe its usage.

Overwriting an item now checks if the item previously contained a user
string.  If so, it recycles its ID for later use.

Completely removed the save logic and icon since saving is automatic.

Attempting to load a project that no longer exists no longer creates it.

Re-wrote handling of operators with respect to item clicking.