Open project now has a bold gold "ACTIVE" next to it in the project list
The default name of a project is now Project # rather than the creation
date.  This makes project names less noisy

Clicking the input element to rename a project no longer selects it

Deleting the last project kicks you back to the editor.
-Also attempts to delete the list database to reset the incrementer

Deleting the active project asks for confirmation and resets editor

Creating a new project closes the menu